+++ .github/CODEOWNERS
@@ -3,5 +3,6 @@
 /relayer @corverroos @jmozah @lazar955
 /lib @corverroos @jmozah @lazar955 @kevinhalliday @DanFlannel @pavelnikolov
 /contracts @kevinhalliday @ttarsi
-/explorerapi @DanFlannel @lazar955
-/indexer @DanFlannel @lazar955
+/explorer @DanFlannel
+/explorer/api @lazar955
+/explorer/indexer @lazar955
