+++ explorer/indexer/main.go
@@ -1,0 +1,11 @@
+// Command api is the main entry point for the api.
+package main
+
+import (
+	appcmd "github.com/omni-network/omni/explorer/indexer/cmd"
+	libcmd "github.com/omni-network/omni/lib/cmd"
+)
+
+func main() {
+	libcmd.Main(appcmd.New())
+}
